Apache web authentication.
For services that I access only with a web browser, using Vouch is fine.
But different clients don't support it, so using basic authentication
backed by LDAP is a better option. So let's try to do that with Apache
with LDAP autheticaion and a similar setup using `auth_request` in
Nginx.