Docker Compose: Small network fix.
There are multiple services deployed by Docker Compose on the host. The
first one creates the Docker network shared by all of them. However,
newer versions of Docker Compose check the network tags and if the tag
for the service doesn't match it (rightfully) complains. Instead the
network is now created by an Ansible playbook when first provisioning
the host and the network is marked as externally managed in all of the
services.